+++ src/datasources/test-plans/TestPlansDataSource.ts
@@ -8,6 +8,7 @@ import { isTimeField, transformDuration } from './utils';
 import { QUERY_TEMPLATES_BATCH_SIZE, QUERY_TEMPLATES_REQUEST_PER_SECOND, QUERY_TEST_PLANS_MAX_TAKE, QUERY_TEST_PLANS_REQUEST_PER_SECOND } from './constants/QueryTestPlans.constants';
 import { AssetUtils } from './asset.utils';
 import { WorkspaceUtils } from 'shared/workspace.utils';
+import { SystemUtils } from 'shared/system.utils';
 import { QueryBuilderOperations } from 'core/query-builder.constants';
 import { ExpressionTransformFunction, transformComputedFieldsQuery } from 'core/query-builder.utils';
 
@@ -20,6 +21,7 @@ export class TestPlansDataSource extends DataSourceBase<TestPlansQuery> {
     super(instanceSettings, backendSrv, templateSrv);
     this.assetUtils = new AssetUtils(instanceSettings, backendSrv);
     this.workspaceUtils = new WorkspaceUtils(this.instanceSettings, this.backendSrv);
+    this.systemUtils = new SystemUtils(instanceSettings, backendSrv);
   }
 
   baseUrl = `${this.instanceSettings.url}/niworkorder/v1`;
@@ -27,6 +29,7 @@ export class TestPlansDataSource extends DataSourceBase<TestPlansQuery> {
   queryTemplatesUrl = `${this.baseUrl}/query-testplan-templates`;
   assetUtils: AssetUtils;
   workspaceUtils: WorkspaceUtils;
+  systemUtils: SystemUtils;
 
   defaultQuery = {
     outputType: OutputType.Properties,
@@ -50,6 +53,7 @@ export class TestPlansDataSource extends DataSourceBase<TestPlansQuery> {
 
   async runQuery(query: TestPlansQuery, options: DataQueryRequest): Promise<DataFrameDTO> {
     const workspaces = await this.workspaceUtils.getWorkspaces();
+    const systemAliases = await this.systemUtils.getSystemAliases();
     if (query.queryBy) {
       query.queryBy = transformComputedFieldsQuery(
         this.templateSrv.replace(query.queryBy, options.scopedVars),
@@ -109,6 +113,9 @@ export class TestPlansDataSource extends DataSourceBase<TestPlansQuery> {
                 return template ? `${template.name} (${template.id})` : value;
               case PropertiesProjectionMap.ESTIMATED_DURATION_IN_SECONDS.label:
                 return value ? transformDuration(value) : '';
+              case PropertiesProjectionMap.SYSTEM_NAME.label:
+                const system = systemAliases.get(value);
+                return system ? system.alias : value;
               case PropertiesProjectionMap.PROPERTIES.label:
                 return value == null ? '' : JSON.stringify(value);
               default: