+++ src/shared/types/QuerySystems.types.ts
@@ -1,0 +1,4 @@
+export interface SystemAlias {
+    id: string;
+    alias: string;
+}